API Design, nothing works

+++ espresso/utilities/Tensor.swift
@@ -13,7 +13,7 @@ import Foundation
  *  Tensor can take arbitrary type of data and when using, should be initialized with the dimension.
  */
 public class Tensor {
-  public typealias DataType = Int
+  public typealias DataType = Float
   
   public var storage : [DataType] = []
